requirements.txt
django 
djangorestframework 
channels 
uvicorn
channels[daphne]
channels_redis
selenium